Move attitude/path interface mail queue definitions
